[PONG:BASE] Classes extracted from types.ts into own files.
