+++ pong/src/utils.ts
@@ -1,4 +1,4 @@
-import { Transform } from "./types";
+import { Transform } from "./transform.js";
 
 const errorMessage = document.getElementById('error-message');
 export function showError(message: string) {
